refactor problema 3
